Add optimal path for JsonElement -> GetInput<T>
